Moved the Home Page to the correct location
